src/requirements.txt
discord.py[voice]==2.2.3